+++ engine/src/evalinfo.cpp
@@ -215,7 +215,7 @@ void update_eval_info(EvalInfo& evalInfo, const Node* rootNode, size_t tbHits, s
     evalInfo.legalMoves = rootNode->get_legal_actions();
 
     vector<size_t> indices;
-    size_t maxIdx = min(searchSettings->multiPV, rootNode->get_no_visit_idx());
+    uint_fast16_t maxIdx = min(uint_fast16_t(searchSettings->multiPV), uint_fast16_t(rootNode->get_no_visit_idx()));
 
     if (maxIdx > 1) {
         sort_eval_lists(evalInfo, indices);
